+++ scripts/control-ui-mock-dev.ts
@@ -16,6 +16,8 @@ import {
   resolveSourcePackageAliasesForVite,
   resolveTsconfigPathAliasesForVite,
 } from "../ui/vite.config.ts";
+import { buildChannelsStatusMock, buildChannelWizardMocks } from "./control-ui-mock-channels.ts";
+import { buildPluginCatalogMock } from "./control-ui-mock-plugins.ts";
 
 type CliOptions = {
   allowedHosts: string[];
@@ -248,191 +250,6 @@ function buildSessionDiffMock() {
     ],
     additions: 6,
     deletions: 2,
-  };
-}
-
-function buildChannelsStatusMock(baseTime: number) {
-  const channelMeta = [
-    { id: "whatsapp", label: "WhatsApp", detailLabel: "WhatsApp Web" },
-    { id: "telegram", label: "Telegram", detailLabel: "Telegram Bot" },
-    { id: "discord", label: "Discord", detailLabel: "Discord Bot" },
-    { id: "slack", label: "Slack", detailLabel: "Slack App" },
-    { id: "signal", label: "Signal", detailLabel: "signal-cli" },
-    { id: "imessage", label: "iMessage", detailLabel: "macOS Messages" },
-    { id: "googlechat", label: "Google Chat", detailLabel: "Chat API" },
-    { id: "nostr", label: "Nostr", detailLabel: "Nostr relays" },
-  ];
-  const account = (params: {
-    accountId?: string;
-    configured?: boolean;
-    running?: boolean;
-    connected?: boolean;
-    lastInboundAt?: number;
-    lastError?: string;
-  }) => ({
-    accountId: params.accountId ?? "default",
-    enabled: params.configured ?? true,
-    configured: params.configured ?? true,
-    running: params.running ?? false,
-    connected: params.connected ?? null,
-    ...(params.lastInboundAt ? { lastInboundAt: params.lastInboundAt } : {}),
-    ...(params.lastError ? { lastError: params.lastError } : {}),
-  });
-  return {
-    ts: baseTime,
-    channelOrder: channelMeta.map((entry) => entry.id),
-    channelLabels: Object.fromEntries(channelMeta.map((entry) => [entry.id, entry.label])),
-    channelDetailLabels: Object.fromEntries(
-      channelMeta.map((entry) => [entry.id, entry.detailLabel]),
-    ),
-    channelMeta,
-    channels: {
-      whatsapp: {
-        configured: true,
-        linked: true,
-        running: true,
-        connected: true,
-        lastConnectedAt: baseTime - 90_000,
-        lastMessageAt: baseTime - 120_000,
-        authAgeMs: 6 * 24 * 60 * 60 * 1000,
-      },
-      telegram: {
-        configured: true,
-        running: true,
-        connected: true,
-        lastStartAt: baseTime - 600_000,
-      },
-      discord: {
-        configured: true,
-        running: false,
-        lastError: "Disallowed intents: enable Message Content in the developer portal.",
-      },
-    },
-    channelAccounts: {
-      whatsapp: [account({ running: true, connected: true, lastInboundAt: baseTime - 120_000 })],
-      telegram: [account({ running: true, connected: true, lastInboundAt: baseTime - 45_000 })],
-      discord: [
-        account({
-          running: false,
-          lastError: "Disallowed intents: enable Message Content in the developer portal.",
-        }),
-      ],
-    },
-    channelDefaultAccountId: {
-      whatsapp: "default",
-      telegram: "default",
-      discord: "default",
-    },
-  };
-}
-
-function buildChannelWizardMocks() {
-  const channelSelectStep = {
-    id: "mock-wizard-step-channel",
-    type: "select",
-    message: "Which channel do you want to set up?",
-    options: [
-      { value: "telegram", label: "Telegram", hint: "bot via @BotFather" },
-      { value: "slack", label: "Slack", hint: "socket mode app" },
-      { value: "signal", label: "Signal", hint: "signal-cli link" },
-      { value: "imessage", label: "iMessage", hint: "macOS Messages" },
-      { value: "__skip__", label: "Skip for now" },
-    ],
-    initialValue: "telegram",
-    executor: "client",
-  };
-  const tokenStep = {
-    id: "mock-wizard-step-token",
-    type: "text",
-    message: "Paste the bot token from @BotFather",
-    placeholder: "paste the token here",
-    sensitive: true,
-    executor: "client",
-  };
-  return {
-    start: {
-      sessionId: "mock-wizard-session",
-      done: false,
-      status: "running",
-      step: channelSelectStep,
-    },
-    next: {
-      cases: [
-        {
-          match: {
-            answer: { stepId: "mock-wizard-step-channel", value: "telegram" },
-          },
-          response: { done: false, status: "running", step: tokenStep },
-        },
-        { response: { done: true, status: "done", channels: ["telegram"] } },
-      ],
-    },
-  };
-}
-
-function buildPluginCatalogMock() {
-  const entry = (params: {
-    id: string;
-    name: string;
-    description: string;
-    category: string;
-    installed: boolean;
-    enabled?: boolean;
-    featured?: boolean;
-  }) => ({
-    id: params.id,
-    name: params.name,
-    description: params.description,
-    version: "1.4.0",
-    installed: params.installed,
-    enabled: params.installed && (params.enabled ?? true),
-    state: params.installed ? ((params.enabled ?? true) ? "enabled" : "disabled") : "not-installed",
-    category: params.category,
-    featured: params.featured ?? false,
-    removable: params.installed,
-  });
-  return {
-    plugins: [
-      entry({
-        id: "telegram",
-        name: "Telegram",
-        description: "Chat with your agent from Telegram DMs and groups.",
-        category: "channel",
-        installed: true,
-      }),
-      entry({
-        id: "discord",
-        name: "Discord",
-        description: "Bridge agents into Discord servers and DMs.",
-        category: "channel",
-        installed: true,
-        enabled: false,
-      }),
-      entry({
-        id: "memory-wiki",
-        name: "Memory Wiki",
-        description: "Long-term wiki-style memory for people and projects.",
-        category: "memory",
-        installed: true,
-      }),
-      entry({
-        id: "browser",
-        name: "Browser",
-        description: "Drive a managed browser profile for research and automation.",
-        category: "tool",
-        installed: false,
-        featured: true,
-      }),
-      entry({
-        id: "canvas",
-        name: "Canvas",
-        description: "Generate and preview visual artifacts from sessions.",
-        category: "tool",
-        installed: false,
-      }),
-    ],
-    diagnostics: [],
-    mutationAllowed: true,
   };
 }
